+++ chevah/compat/nt_filesystem.py
@@ -167,6 +167,7 @@ class NTFilesystem(PosixFilesystemBase):
         * unlock
           * [path1] -> path1:\
           * [path1, path2] -> path1 :\ path2
+          * [UNC, server1, path1, path2] -> \\server1\path1\path2
         '''
         if segments is None or len(segments) == 0:
             result = self._root_path
@@ -175,18 +176,23 @@ class NTFilesystem(PosixFilesystemBase):
         else:
             drive = u'%s:\\' % segments[0]
             path_segments = segments[1:]
+
             if len(path_segments) == 0:
                 result = drive
             else:
-                result = os.path.normpath(
-                    drive + os.path.join(*path_segments))
-                # os.path.normpath can result in an 'out of drive' path.
-                if result.find(':\\') == -1:
-                    if result.find('\\') == -1:
-                        result = result + ':\\'
-                    else:
-                        result = result.replace('\\', ':\\', 1)
-            self._validateDrivePath(result)
+                if drive == u'UNC:\\':
+                    result = '\\' + os.path.normpath(
+                        '\\' + os.path.join(*path_segments))
+                else:
+                    result = os.path.normpath(
+                        drive + os.path.join(*path_segments))
+                    # os.path.normpath can result in an 'out of drive' path.
+                    if result.find(':\\') == -1:
+                        if result.find('\\') == -1:
+                            result = result + ':\\'
+                        else:
+                            result = result.replace('\\', ':\\', 1)
+        self._validateDrivePath(result)
 
         return text_type(result)
 
@@ -201,6 +207,11 @@ class NTFilesystem(PosixFilesystemBase):
         """
         Raise an error if path does not have valid driver.
         """
+        if path.startswith('\\\\'):
+            # We have a network path and we don't check the server's
+            # availability.
+            return
+
         path_encoded = self.getEncodedPath(path)
         letter, _ = os.path.splitdrive(path_encoded)
         if letter.strip(':').lower() not in self._allowed_drive_letters:
@@ -218,6 +229,8 @@ class NTFilesystem(PosixFilesystemBase):
             return segments
 
         head = True
+
+        original_path = path
         path = os.path.abspath(path)
 
         if self._avatar.lock_in_home_folder:
@@ -226,13 +239,22 @@ class NTFilesystem(PosixFilesystemBase):
             tail = path[len(self._getRootPath()):]
             drive = ''
         else:
-            # For unlocked filesystem, we use 'c' as default drive.
-            drive, root_tail = os.path.splitdrive(path)
-            if not drive:
-                drive = u'c'
+            if original_path.startswith('UNC\\'):
+                # We have a network path.
+                drive = u'UNC'
+                tail = os.path.abspath(original_path[3:])
+            elif original_path.startswith('\\\\?\\UNC\\'):
+                # We have a long UNC.
+                drive = u'UNC'
+                tail = os.path.abspath(original_path[7:])
             else:
-                drive = drive.strip(u':')
-            tail = root_tail
+                # For unlocked filesystem, we use 'c' as default drive.
+                drive, root_tail = os.path.splitdrive(path)
+                if not drive:
+                    drive = u'c'
+                else:
+                    drive = drive.strip(u':')
+                tail = root_tail
 
         while head not in [u'\\', u'']:
             head, tail = os.path.split(tail)
@@ -278,7 +300,13 @@ class NTFilesystem(PosixFilesystemBase):
         Tries to mimic behaviour of Unix readlink command.
         """
         path = self.getRealPathFromSegments(segments)
+        result = self._readLink(path)
+        return self.getSegmentsFromRealPath(result['target'])
 
+    def _readLink(self, path):
+        """
+        Return a dict with the link target.
+        """
         try:
             handle = win32file.CreateFileW(
                 path,
@@ -316,7 +344,7 @@ class NTFilesystem(PosixFilesystemBase):
             message = u'%s %s' % (error.message, path)
             raise OSError(errno.EINVAL, message.encode('utf-8'))
 
-        return self.getSegmentsFromRealPath(result['target'])
+        return result
 
     def makeLink(self, target_segments, link_segments):
         """
@@ -330,12 +358,13 @@ class NTFilesystem(PosixFilesystemBase):
         # FIXME:2025:
         # Add support for junctions.
         if not self.process_capabilities.symbolic_link:
-            raise NotImplementedError
+            raise NotImplementedError('makeLink not implemented on this OS.')
 
         target_path = self.getRealPathFromSegments(target_segments)
         link_path = self.getRealPathFromSegments(link_segments)
 
-        if self.isFolder(target_segments):
+        if self.isFolder(target_segments) or target_path.startswith('\\UNC\\'):
+            # We have folder or a Windows share as target.
             flags = win32file.SYMBOLIC_LINK_FLAG_DIRECTORY
         else:
             flags = 0
@@ -384,22 +413,6 @@ class NTFilesystem(PosixFilesystemBase):
         # Set node_id by concatenating the volume and the file_id.
         stats_list[1] = volume_id << 64 | index_high << 32 | index_low
         return type(stats)(stats_list)
-
-    def getAttributes(self, segments):
-        """
-        See `ILocalFilesystem`.
-        """
-        if not self.exists(segments):
-            # On Windows, it will return the attributes, even if the target
-            # does not exists.
-            raise OSError(
-                errno.ENOENT,
-                'No such file or directory',
-                self.getRealPathFromSegments(segments),
-                )
-
-        with self._windowsToOSError(segments):
-            return super(NTFilesystem, self).getAttributes(segments)
 
     def setAttributes(self, segments, attributes):
         '''See `ILocalFilesystem`.'''
@@ -468,7 +481,7 @@ class NTFilesystem(PosixFilesystemBase):
             return super(NTFilesystem, self).createFolder(
                 segments, recursive)
 
-    def _getFileData(self, segments):
+    def _getFileData(self, path):
         """
         Return a dict with resolved WIN32_FIND_DATA structure for segments.
 
@@ -478,8 +491,6 @@ class NTFilesystem(PosixFilesystemBase):
         http://msdn.microsoft.com/en-us/library/windows/
             desktop/gg258117(v=vs.85).aspx
         """
-        path = self.getEncodedPath(self.getRealPathFromSegments(segments))
-
         try:
             with self._impersonateUser():
                 search = win32file.FindFilesW(path)
@@ -534,13 +545,20 @@ class NTFilesystem(PosixFilesystemBase):
         See `ILocalFilesystem`.
         """
         try:
-            data = self._getFileData(segments)
-            is_reparse_point = bool(
-                data['attributes'] & FILE_ATTRIBUTE_REPARSE_POINT)
-            has_symlink_tag = (data['tag'] == self.IO_REPARSE_TAG_SYMLINK)
-            return is_reparse_point and has_symlink_tag
+            path = self.getRealPathFromSegments(segments)
+            return self._isLink(path)
         except OSError:
             return False
+
+    def _isLink(self, path):
+        """
+        Return True if path is a symlink.
+        """
+        data = self._getFileData(self.getEncodedPath(path))
+        is_reparse_point = bool(
+            data['attributes'] & FILE_ATTRIBUTE_REPARSE_POINT)
+        has_symlink_tag = (data['tag'] == self.IO_REPARSE_TAG_SYMLINK)
+        return is_reparse_point and has_symlink_tag
 
     def deleteFile(self, segments, ignore_errors=False):
         """
@@ -789,16 +807,3 @@ class NTFilesystem(PosixFilesystemBase):
                 if group_sid == sid:
                     return True
         return False
-
-    def exists(self, segments):
-        """
-        See `ILocalFilesystem`.
-        """
-        if self.isLink(segments):
-            try:
-                target_segments = self.readLink(segments)
-                return self.exists(target_segments)
-            except CompatError:
-                return False
-        else:
-            return super(NTFilesystem, self).exists(segments)