+++ pavement.py
@@ -86,7 +86,7 @@ BUILD_PACKAGES = [
 
 # Packages required by the static analysis tests.
 LINT_PACKAGES = [
-    'scame==0.3.3',
+    'scame==0.4.2',
     'pyflakes==1.5.0',
     'pycodestyle==2.3.1',
     'bandit==1.4.0',
@@ -109,7 +109,6 @@ TEST_PACKAGES = [
     'mock',
 
     'coverage==4.4.1',
-    'coverator==0.1.0',
 
     # used for remote debugging.
     'remote_pdb==1.2.0',
@@ -155,7 +154,7 @@ test_super
 try:
     from scame.formatcheck import ScameOptions
 
-    class ServerScameOptions(ScameOptions):
+    class CompatScameOptions(ScameOptions):
         """
         Scame options for the this project.
         """
@@ -178,8 +177,9 @@ try:
 
             return getattr(self, option)
 
-    options = ServerScameOptions()
+    options = CompatScameOptions()
     options.max_line_length = 80
+    options.progress = True
 
     options.scope = {
         'include': [