+++ release-notes.rst
@@ -1,6 +1,19 @@
 Release notes for chevah.compat
 ===============================
 
+
+0.49.1 - 02/04/2018
+-------------------
+
+* ILocalFilesystem.exist no longer follows links.
+
+
+0.49.0 - 30/04/2018
+-------------------
+
+* Add support for working with UNC paths and symbolic links to Windows shares.
+
+
 0.48.0 - 15/04/2018
 -------------------